refactor, extract hydra related variables and methods to HydraUtil
