+++ obp-api/pom.xml
@@ -87,7 +87,7 @@
     <dependency>
       <groupId>org.bouncycastle</groupId>
       <artifactId>bcpg-jdk15on</artifactId>
-      <version>1.47</version>
+      <version>1.62</version>
     </dependency>
     <dependency>
       <groupId>org.bouncycastle</groupId>
